refactor(game): game 패키지 생성 및 Computer 클래스 패키지 이전
- game 패키지는 게임 관련 로직 및 기능을 구현하는 클래스들의 집합
