+++ src/main/java/baseball/Player.java
@@ -1,6 +1,7 @@
 package baseball;
 
 import constants.StringConstants;
+import game.Computer;
 
 public class Player {
